% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiktokads-package.R, R/rtiktokads_package.R
\docType{package}
\name{rtiktokads-package}
\alias{rtiktokads-package}
\title{rtiktokads: Work with [TikTok Marketing API]}
\description{
Loading dtaa from [TikTok Marketing API] by bussines centers, advertisers, budgets and reports.

Loading dtaa from [TikTok Marketing API] by bussines centers, advertisers, budgets and reports.
}
\author{
\strong{Maintainer}: Alexey Seleznev \email{selesnow@gmail.com} (\href{https://orcid.org/0000-0003-0410-7385}{ORCID})

Other contributors:
\itemize{
  \item Netpeak [copyright holder]
}

}
\keyword{internal}
