% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tik_auth.R
\name{tik_auth}
\alias{tik_auth}
\title{Authorization in 'TikTok Marketing API'}
\usage{
tik_auth(username = tik_get_username(), token_path = tik_get_token_path())
}
\arguments{
\item{username}{TikTok login}

\item{token_path}{auth cache folder}
}
\value{
tik_access_token object
}
\description{
Authorization in 'TikTok Marketing API'
}
