% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tik_get_ad_groups.R
\name{tik_get_ad_groups}
\alias{tik_get_ad_groups}
\title{Get ad groups}
\usage{
tik_get_ad_groups(advertiser_id, fields = NULL)
}
\arguments{
\item{advertiser_id}{Advertiser ID.}

\item{fields}{Fields that you want to get.}
}
\description{
Get ad groups
}
