% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tik_get_ads.R
\name{tik_get_ads}
\alias{tik_get_ads}
\title{Get ads}
\usage{
tik_get_ads(advertiser_id, fields = NULL)
}
\arguments{
\item{advertiser_id}{Advertiser ID.}

\item{fields}{Fields that you want to get.}
}
\description{
Get ads
}
