% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tik_get_advertiser_accounts.R
\name{tik_get_advertiser_accounts}
\alias{tik_get_advertiser_accounts}
\title{Get authorized ad accounts}
\usage{
tik_get_advertiser_accounts()
}
\value{
tibble with advertisers name and id
}
\description{
Get authorized ad accounts
}
\examples{
\dontrun{
advertisers <- tik_get_advertiser_accounts()
}
}
