% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tik_get_advertiser_info.R
\name{tik_get_advertiser_info}
\alias{tik_get_advertiser_info}
\title{Get ad account details}
\usage{
tik_get_advertiser_info(advertiser_ids, fields = NULL)
}
\arguments{
\item{advertiser_ids}{List of advertiser IDs to query. You can obtain Advertiser IDs through the \code{tik_get_advertiser_accounts()} function}

\item{fields}{A list of information to be returned. Supported values:
\itemize{
\item telephone_number
\item contacter
\item currency
\item cellphone_number
\item timezone
\item advertiser_id
\item role
\item company
\item status
\item description
\item rejection_reason
\item address
\item name
\item language
\item industry
\item license_no
\item email
\item license_url
\item country
\item balance
\item create_time
\item display_timezone
\item owner_bc_id
\item company_name_editable.
}}
}
\value{
tibble with advertisers onfo
}
\description{
Get ad account details
}
\examples{
\dontrun{
advertisers <- tik_get_advertiser_accounts()
advertisers_info <- tik_get_advertiser_info(advertisers$advertiser_id)
}
}
