% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tik_get_bc_partners.R
\name{tik_get_bc_partners}
\alias{tik_get_bc_partners}
\title{Get the partners of a BC}
\usage{
tik_get_bc_partners(bc_id)
}
\arguments{
\item{bc_id}{Business Center ID.}
}
\value{
tibble with list of partners
}
\description{
Get the partners of a BC
}
