% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tik_get_business_centers.R
\name{tik_get_business_centers}
\alias{tik_get_business_centers}
\title{Get Business Centers}
\usage{
tik_get_business_centers()
}
\value{
tibble with business centers info
}
\description{
Get Business Centers
}
\examples{
\dontrun{
bc <- tik_get_business_centers()
}
}
