% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tik_get_campaigns.R
\name{tik_get_campaigns}
\alias{tik_get_campaigns}
\title{Get campaigns}
\usage{
tik_get_campaigns(advertiser_id, fields = NULL)
}
\arguments{
\item{advertiser_id}{Advertiser ID.}

\item{fields}{Fields that you want to get.}
}
\description{
Get campaigns
}
