% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tik_auth.R
\name{tik_set_token_path}
\alias{tik_set_token_path}
\title{Change session token path}
\usage{
tik_set_token_path(
  token_path = ifelse(getOption("tiktok.token_path") == "global",
    rappdirs::site_data_dir("rtiktokads"), rappdirs::user_cache_dir("rtiktokads"))
)
}
\arguments{
\item{token_path}{auth cache folder}
}
\value{
using for side effect, no return value
}
\description{
Change session token path
}
