% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tik_auth.R
\name{tik_set_username}
\alias{tik_set_username}
\title{Set session auth username}
\usage{
tik_set_username(username)
}
\arguments{
\item{username}{TikTok login}
}
\value{
using for side effect, no return value
}
\description{
Set session auth username
}
